requirements.txt
Django
streamlit
googletrans
djangorestframework
django-cors-headers
googletrans
diffusers
transformers
torch
numpy
pyyaml
pandas
matplotlib
scipy
scikit-learn
Pillow
psycopg2-binary
mysqlclient